+++ src/RoleCard/RoleCard.tsx
@@ -6,7 +6,12 @@ export default function RoleCard({ role, link }) {
     // <Card className='roleCard' shadow="lg" radius="xl" p="xl" style={{ width: 'calc(33.333% - 20px)', marginBottom: '20px' }}>
     <Card className="roleCard" shadow="lg" radius="xl" p="xl">
       <Center>
-        <Image className="moniter" src="/monitor.svg" />
+        <Image
+          className="moniter"
+          src="/monitor.svg"
+          width="4vw"
+          height="4vw"
+        />
       </Center>
       <Center my="xs">
         <Text c="#0078E8" fw={500} className="roleTitle">
@@ -15,7 +20,14 @@ export default function RoleCard({ role, link }) {
       </Center>
       <Center>
         <Button
-          rightSection={<Image className="chevron" src="/chevron-right.svg" />}
+          rightSection={
+            <Image
+              className="chevron"
+              src="/chevron-right.svg"
+              width="1.5vw"
+              height="1.5vw"
+            />
+          }
           variant="filled"
           radius="xl"
           size="xl"
