+++ src/for-students-FAQ/FaqCard.css
