+++ src/for-students-FAQ/FaqCard.tsx
